requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.4
Jinja2==2.7.2
MarkupSafe==0.23
SQLAlchemy==0.7.9
Tempita==0.5.2
WTForms==2.0
Werkzeug==0.9.4
argparse==1.2.1
beautifulsoup4==4.3.2
decorator==3.4.0
feedparser==5.1.3
gunicorn==18.0
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
psycopg2==2.5